Move arrow graph constructor into its own file
