+++ graphdatascience/query_runner/arrow_graph_constructor.py
@@ -1,0 +1,104 @@
+import json
+import math
+from concurrent.futures import ThreadPoolExecutor, wait
+from typing import Dict, List
+
+import pyarrow.flight as flight
+from pandas.core.frame import DataFrame
+from pyarrow import Table
+
+from .graph_constructor import GraphConstructor
+from .query_runner import QueryRunner
+
+
+class ArrowGraphConstructor(GraphConstructor):
+    def __init__(
+        self,
+        query_runner: QueryRunner,
+        graph_name: str,
+        flight_client: flight.FlightClient,
+        flight_options: flight.FlightCallOptions,
+        concurrency: int,
+        chunk_size: int = 10_000,
+    ):
+        self._query_runner = query_runner
+        self._concurrency = concurrency
+        self._graph_name = graph_name
+        self._client = flight_client
+        self._flight_options = flight_options
+        self._chunk_size = chunk_size
+        self._min_batch_size = chunk_size * 10
+
+    def run(self, node_dfs: List[DataFrame], relationship_dfs: List[DataFrame]) -> None:
+        try:
+            self._send_action(
+                "CREATE_GRAPH", {"name": self._graph_name, "database_name": self._query_runner.database()}
+            )
+
+            partitioned_node_dfs = self._partition_dfs(node_dfs)
+            with ThreadPoolExecutor(self._concurrency) as executor:
+                futures = [executor.submit(self._send_df, df, "node") for df in partitioned_node_dfs]
+
+                wait(futures)
+
+                for future in futures:
+                    if not future.exception():
+                        continue
+                    raise future.exception()  # type: ignore
+
+            self._send_action("NODE_LOAD_DONE", {"name": self._graph_name})
+
+            partitioned_rel_dfs = self._partition_dfs(relationship_dfs)
+            with ThreadPoolExecutor(self._concurrency) as executor:
+                futures = [executor.submit(self._send_df, df, "relationship") for df in partitioned_rel_dfs]
+
+                wait(futures)
+
+                for future in futures:
+                    if not future.exception():
+                        continue
+                    raise future.exception()  # type: ignore
+
+            self._send_action("RELATIONSHIP_LOAD_DONE", {"name": self._graph_name})
+        except Exception as e:
+            self._send_action("ABORT", {"name": self._graph_name})
+            raise e
+
+    def _partition_dfs(self, dfs: List[DataFrame]) -> List[DataFrame]:
+        partitioned_dfs = []
+
+        for df in dfs:
+            num_rows = df.shape[0]
+
+            if num_rows <= self._min_batch_size:
+                partitioned_dfs.append(df)
+                continue
+
+            num_batches = math.ceil(num_rows / self._min_batch_size)
+            for i in range(num_batches):
+                row_offset = i * self._min_batch_size
+                partitioned_dfs.append(
+                    df.iloc[row_offset : min(num_rows, row_offset + self._min_batch_size), :]  # noqa: E203
+                )
+
+        return partitioned_dfs
+
+    def _send_action(self, action_type: str, meta_data: Dict[str, str]) -> None:
+        result = self._client.do_action(
+            flight.Action(action_type, json.dumps(meta_data).encode("utf-8")), self._flight_options
+        )
+
+        json.loads(next(result).body.to_pybytes().decode())
+
+    def _send_df(self, df: DataFrame, entity_type: str) -> None:
+        table = Table.from_pandas(df)
+        flight_descriptor = {"name": self._graph_name, "entity_type": entity_type}
+
+        # Write schema
+        upload_descriptor = flight.FlightDescriptor.for_command(json.dumps(flight_descriptor).encode("utf-8"))
+
+        writer, _ = self._client.do_put(upload_descriptor, table.schema, self._flight_options)
+
+        with writer:
+            # Write table in chunks
+            writer.write_table(table, max_chunksize=self._chunk_size)
